PR feedback: Get rid of release_1_9 package; Enable for CircleCI; get rid of Travis Use Java 11
