--- .travis.yml
@@ -1,2 +1,0 @@
-language: java
-jdk: openjdk8
