+++ pom.xml
@@ -7,8 +7,8 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>1.11</maven.compiler.source>
+        <maven.compiler.target>1.11</maven.compiler.target>
         <scala.version>2.13.7</scala.version>
         <encoding>UTF-8</encoding>
         <gatling-plugin.version>3.0.0</gatling-plugin.version>
